AQ-46 Refactored code.
