+++ quiz/web/src/components/TestQuery/api.ts
@@ -1,19 +1,6 @@
 import { useQuery, UseQueryResult } from "@tanstack/react-query";
 import axios from "axios";
 
-export default function TestQuery() {
-    const { data, isLoading } = useCatFact();
-
-    if (isLoading || !data) {
-        return <h2>Loading</h2>;
-    }
-    return (
-        <div>
-            <h2>{data.fact}</h2>
-        </div>
-    );
-}
-
 interface CatFact {
     fact: string;
     length: number;
@@ -24,7 +11,7 @@ async function fetchCatFact(): Promise<CatFact> {
     return response.data;
 }
 
-function useCatFact(): UseQueryResult<CatFact> {
+export default function useCatFact(): UseQueryResult<CatFact> {
     return useQuery({
         queryKey: ["catFact"],
         queryFn: fetchCatFact,
