+++ quiz/web/src/components/TestQuery/index.ts
@@ -1,0 +1,3 @@
+import TestQuery from "components/TestQuery/TestQuery";
+
+export default TestQuery;
